+++ src/common/globals.ts
@@ -1,1 +1,8 @@
-export const origins = ["http://localhost:5173", "http://localhost:5174"];
+export const origins = [
+  "http://localhost:5173",
+  "http://localhost:5174",
+  "https://jamesmdrakemedia.com",
+  "https://jamesdraketech.com",
+  "https://beardonbass.com",
+  "https://rides.jamesdraketech.com",
+];